rag-ingest/requirements.txt
pypdf
python-docx
requests
qdrant-client==1.9.0
python-dotenv
openai
